shl_real/requirements.txt
fastapi
uvicorn
faiss-cpu
sentence-transformers
numpy
pandas
openpyxl
